clarity-models/requirements.txt
transformers[torch]==4.57.1
torch==2.9.0
datasets==2.21.0
fastapi==0.119.1
uvicorn[standard]==0.38.0
accelerate>=0.21.0
loguru==0.7.3
peft==0.17.1
PyYAML==6.0.3
bitsandbytes==0.42.0
sentencepiece==0.2.1
tensorboard==2.20.0